Finished irMov function and added function to convert char[] to a hex value -getHexValue(char ptr[], int bitcount)-

+++ registers_codes.h
@@ -2,21 +2,21 @@
 #define REGISTERS_CODES_H_
 
 // Registers
-int rax;
-int rcx;
-int rdx;
-int rbx;
-int rsp;
-int rbp;
-int rsi;
-int rdi;
-int r8;
-int r9;
-int r10;
-int r11;
-int r12;
-int r13;
-int r14;
+long int rax;
+long int rcx;
+long int rdx;
+long int rbx;
+long int rsp;
+long int rbp;
+long int rsi;
+long int rdi;
+long int r8;
+long int r9;
+long int r10;
+long int r11;
+long int r12;
+long int r13;
+long int r14;
 
 
 int of; // Overflow flag
